# math/big: optimize amd64 asm shlVU and shrVU for shift==0 case